requirements.txt
numpy==1.24.3
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.1.3
torch==1.13.0+cu117
